services/transcribe-api/requirements.txt
fastapi
uvicorn[standard]
openai-whisper
torch  # with CUDA for GPU
pyannote.audio
soundfile
